refactor: pass button information to children
